Created the AnalyzeModel static class, moved analysis functions.
